pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refrigerant-optimization"
version = "0.1.0"
description = "Linear optimization model for refrigerant recycling and distillation"
authors = [
    {name = "Daksh Babbar", email = "daksh@dakshbabbar.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pulp>=2.7.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/daksh-7/refrigerant-optimization-chemours"
Repository = "https://github.com/daksh-7/refrigerant-optimization-chemours"
Issues = "https://github.com/daksh-7/refrigerant-optimization-chemours/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
refrigerant-opt = "src.cli:main"

[tool.setuptools]
packages = ["src"]